src/redpanda-init/requirements.txt
kafka-python
PyYAML==6.0.1
